# FrontEnd/OpenAPIs/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "BackEnd",
    "version": "1.0"
  },
  "paths": {
    "/rolelist": {
      "get": {
        "tags": [
          "Roles"
        ],
        "operationId": "GetRoleList",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/GroupDto"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/rolelistddl": {
      "get": {
        "tags": [
          "Roles"
        ],
        "operationId": "GetRoleListDDL",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/GroupDdl"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/role/add": {
      "post": {
        "tags": [
          "Roles"
        ],
        "operationId": "AddRole",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RoleDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/role/edit/{roleId}": {
      "put": {
        "tags": [
          "Roles"
        ],
        "operationId": "UpdateRole",
        "parameters": [
          {
            "name": "roleId",
            "in": "path",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RoleDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/role/delete/{id}": {
      "delete": {
        "tags": [
          "Roles"
        ],
        "operationId": "DeleteRole",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/userlist": {
      "get": {
        "tags": [
          "Users"
        ],
        "operationId": "UserList",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/UserList"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/userlistaerverside": {
      "post": {
        "tags": [
          "Users"
        ],
        "operationId": "GetUserListServersided",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GridServerSide"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Int32UserDtoList<>f__AnonymousType10"
                }
              }
            }
          }
        }
      }
    },
    "/user/add": {
      "post": {
        "tags": [
          "Users"
        ],
        "operationId": "AddUser",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/user/edit/{userId}": {
      "put": {
        "tags": [
          "Users"
        ],
        "operationId": "UpdateUser",
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/user/delete/{userId}": {
      "delete": {
        "tags": [
          "Users"
        ],
        "operationId": "DeleteUser",
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/userandroles": {
      "get": {
        "tags": [
          "Users"
        ],
        "operationId": "GetUserAndRoles",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GroupRoleDdl"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "GridServerSide": {
        "type": "object",
        "properties": {
          "skip": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "take": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "requireTotalCount": {
            "type": "boolean",
            "nullable": true
          },
          "sort": {
            "type": "string",
            "nullable": true
          },
          "filter": {
            "type": "array",
            "items": { },
            "nullable": true
          },
          "sorting": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SortingInfo"
            },
            "nullable": true,
            "readOnly": true
          }
        },
        "additionalProperties": false
      },
      "GroupDdl": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "text": {
            "maxLength": 255,
            "type": "string",
            "nullable": true
          },
          "extend": {
            "maxLength": 255,
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "GroupDto": {
        "type": "object",
        "properties": {
          "roleId": {
            "type": "integer",
            "format": "int32"
          },
          "groupId": {
            "type": "integer",
            "format": "int32"
          },
          "appId": {
            "maxLength": 20,
            "type": "string",
            "nullable": true
          },
          "groupName": {
            "maxLength": 1000,
            "type": "string",
            "nullable": true
          },
          "group": {
            "maxLength": 255,
            "type": "string",
            "nullable": true
          },
          "roleName": {
            "maxLength": 50,
            "type": "string",
            "nullable": true
          },
          "isDelete": {
            "type": "boolean"
          }
        },
        "additionalProperties": false
      },
      "GroupRoleDdl": {
        "type": "object",
        "properties": {
          "groups": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/GroupDdl"
            },
            "nullable": true
          },
          "roles": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/GroupDdl"
            },
            "nullable": true
          },
          "package": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/GroupDdl"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "Int32UserDtoList<>f__AnonymousType10": {
        "type": "object",
        "properties": {
          "totalCount": {
            "type": "integer",
            "format": "int32"
          },
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/UserDto"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "RoleDto": {
        "type": "object",
        "properties": {
          "roleId": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "roleName": {
            "type": "string",
            "nullable": true
          },
          "groupId": {
            "type": "integer",
            "format": "int32"
          },
          "groupName": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "SortingInfo": {
        "type": "object",
        "properties": {
          "selector": {
            "type": "string",
            "nullable": true
          },
          "desc": {
            "type": "boolean"
          }
        },
        "additionalProperties": false
      },
      "UserDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "userName": {
            "type": "string",
            "nullable": true
          },
          "pass": {
            "type": "string",
            "nullable": true
          },
          "namaLengkap": {
            "type": "string",
            "nullable": true
          },
          "email": {
            "type": "string",
            "nullable": true
          },
          "alamat": {
            "type": "string",
            "nullable": true
          },
          "noKtp": {
            "type": "integer",
            "format": "int64",
            "nullable": true
          },
          "roles": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "nullable": true
          },
          "users": {
            "type": "array",
            "items": {
              "type": "integer",
              "format": "int32"
            },
            "nullable": true
          },
          "userIds": {
            "type": "array",
            "items": {
              "type": "integer",
              "format": "int32"
            },
            "nullable": true,
            "readOnly": true
          }
        },
        "additionalProperties": false
      },
      "UserList": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "userName": {
            "type": "string",
            "nullable": true
          },
          "email": {
            "type": "string",
            "nullable": true
          },
          "alamat": {
            "type": "string",
            "nullable": true
          },
          "noKtp": {
            "type": "integer",
            "format": "int64",
            "nullable": true
          },
          "namaLengkap": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      }
    }
  }
}